+++ deploy/ansible/roles/metrics/files/grafana/provisioning/dashboards/slot_metrics.json
