+++ packages/app/jsonmodels/slots.go
