+++ packages/core/memstorage/slotstorage.go
