+++ plugins/dashboard/frontend/src/app/components/ExplorerSlotQueryResult.tsx
