+++ plugins/dashboard/frontend/src/app/stores/SlotStore.tsx
