+++ plugins/metrics/metrics_slots.go
