+++ tools/docker-network/grafana/dashboards/slot_metrics.json
